small refactors, removed lodash as dep, added react and react dom as dev deps, fixed onload behaviour
