+++ demo/demo.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import domready from 'domready';
-import Popout from '../lib/react-popout.jsx';
+import Popout from '../lib/react-popout';
 
 class Example extends React.Component {
   constructor(props) {
